3-array_range.c

+++ 0x0C-more_malloc_free/3-array_range.c
@@ -16,7 +16,7 @@ int *array_range(int min, int max)
 	if (min > max)
 		return (NULL);
 
-	size = (unsigned int)(max - min) + 1;
+	size = (max - min) + 1;
 
 	a = malloc(size * sizeof(int));
